+++ Sources/VCore/Services and Managers/User Defaults Service/UserDefaultsServiceError.swift
@@ -1,33 +1,33 @@
 //
-//  KeychainServiceError.swift
+//  UserDefaultsServiceError.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 23.07.22.
+//  Created by Vakhtang Kontridze on 08.09.24.
 //
 
 import Foundation
 
-// MARK: - Keychain Service Error
-/// An error that occurs during the operations in `KeychainService`.
-public struct KeychainServiceError: VCoreError, Equatable {
+// MARK: - User Defaults Service Error
+/// An error that occurs during the operations in `UserDefaultsService`.
+public struct UserDefaultsServiceError: VCoreError, Equatable {
     // MARK: Properties
     private let code: Code
-    
+
     // MARK: Initializers
-    /// Initializes `KeychainServiceError` with the given error code.
+    /// Initializes `UserDefaultsServiceError` with the given error code.
     public init(_ code: Code) {
         self.code = code
     }
-    
+
     // MARK: Code
     /// Error code.
     public enum Code: Int, Equatable {
         /// Indicates that get operation has failed.
         case failedToGet
-        
+
         /// Indicates that set operation has failed.
         case failedToSet
-        
+
         /// Indicates that delete operation has failed.
         case failedToDelete
     }
@@ -37,9 +37,9 @@ public struct KeychainServiceError: VCoreError, Equatable {
 
     public var vCoreErrorDescription: String {
         switch code {
-        case .failedToGet: "Data cannot be retrieved from 'Security' framework"
-        case .failedToSet: "Data cannot be set to 'Security' framework"
-        case .failedToDelete: "Data cannot be deleted from 'Security' framework"
+        case .failedToGet: "Data cannot be retrieved from 'UserDefaults'"
+        case .failedToSet: "Data cannot be set to 'UserDefaults'"
+        case .failedToDelete: "Data cannot be deleted from 'UserDefaults'"
         }
     }